Now using strings as dict keys throughout.
